requirements.txt
# WhatsApp API library (example: Twilio)
twilio

# Linting
flake8
black

# Testing
pytest

# Environment variable management
python-dotenv
